requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
msal==1.30.0
requests==2.32.3
openpyxl==3.1.5
python-dotenv==1.0.0
bleach==6.1.0
PyJWT==2.8.0
Werkzeug==3.0.3
